setup working with packages imports for operations
