+++ erpnext/patches/v7_1/update_lead_source.py
@@ -2,7 +2,7 @@ import frappe
 from frappe import _
 
 def execute():
-	from erpnext.setup.setup_wizard.install_fixtures import default_lead_sources
+	from erpnext.setup.setup_wizard.operations.install_fixtures import default_lead_sources
 
 	frappe.reload_doc('selling', 'doctype', 'lead_source')
