+++ erpnext/setup/setup_wizard/data/industry_type.py
