+++ erpnext/setup/setup_wizard/operations/default_website.py
